Improves the release script and docker build caching
